--- lib/shopify_transporter/exporter.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-require_relative 'exporter/exporter'
-require_relative 'exporter/magento'
