--- lib/shopify_transporter/exporter/magento/soap.rb
@@ -1,41 +1,0 @@
-require 'json'
-require 'savon'
-
-module ShopifyTransporter
-  class Soap
-    def initialize(hostname, username, api_key)
-      @hostname = hostname
-      @username = username
-      @api_key = api_key
-    end
-
-    def call(fn, params)
-      soap_client.call(
-        fn,
-        message: { session_id: soap_session_id }.merge(params)
-      )
-    end
-
-    private
-
-    attr_accessor :hostname, :username, :api_key
-
-    def soap_client
-      @soap_client ||= Savon.client(
-        wsdl: "https://#{hostname}/api/v2_soap?wsdl",
-        open_timeout: 500,
-        read_timeout: 500,
-      )
-    end
-
-    def soap_session_id
-      @soap_session_id ||= soap_client.call(
-        :login,
-        message: {
-          username: username,
-          api_key: api_key,
-        }
-      ).body[:login_response][:login_return]
-    end
-  end
-end
