+++ lib/shopify_transporter/exporters.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+require_relative 'exporters/exporter'
+require_relative 'exporters/magento'
