+++ lib/shopify_transporter/exporters/magento.rb
