+++ spec/shopify_transporter/exporters/magento/magento_order_exporter_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 require 'shopify_transporter/pipeline/stage'
 
-RSpec.describe ShopifyTransporter::MagentoOrderExporter do
+RSpec.describe ShopifyTransporter::Exporters::Magento::MagentoOrderExporter do
   context '#run' do
     it 'retrieves orders from Magento using the SOAP API and returns the results' do
       soap_client = double("soap client")
@@ -43,7 +43,7 @@ RSpec.describe ShopifyTransporter::MagentoOrderExporter do
         }
       ]
 
-       expect(ShopifyTransporter::MagentoOrderExporter.new(1, soap_client).export).to eq(expected_result)
+       expect(ShopifyTransporter::Exporters::Magento::MagentoOrderExporter.new(1, soap_client).export).to eq(expected_result)
     end
   end
 end
